[DROOLS-6473] Refactor ConcurrencyTest in drools-test-coverage - Moved to a new package - Refactor
