+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/concurrency/ConsequenceConcurrencyTest.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.compiler.integrationtests.concurrency;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
+import org.drools.testcoverage.common.util.KieBaseTestConfiguration;
+import org.drools.testcoverage.common.util.TestParametersUtil;
+import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.kie.api.runtime.KieSession;
+import org.kie.test.testcategory.TurtleTestCategory;
+
+@RunWith(Parameterized.class)
+@Category(TurtleTestCategory.class)
+public class ConsequenceConcurrencyTest extends BaseConcurrencyTest {
+
+    public ConsequenceConcurrencyTest(final KieBaseTestConfiguration kieBaseTestConfiguration) {
+        super(kieBaseTestConfiguration);
+    }
+
+    @Parameterized.Parameters(name = "KieBase type={0}")
+    public static Collection<Object[]> getParameters() {
+        return TestParametersUtil.getKieBaseCloudConfigurations(true);
+    }
+
+    protected String getDrl() {
+        return "package com.example.reproducer\n" +
+               "import " + Bus.class.getCanonicalName() + ";\n" +
+               "dialect \"mvel\"\n" +
+               "global java.util.List result;\n" +
+               "rule \"rule_mt_1a\"\n" +
+               "    when\n" +
+               "        $bus : Bus( $title: \"POWER PLANT\" )\n" +
+               "    then\n" +
+               "        result.add($bus.karaoke.dvd[$title].artist);\n" +
+               "end";
+    }
+
+    protected void setGlobal(KieSession kSession) {
+        List<String> result = new ArrayList<>();
+        kSession.setGlobal("result", result);
+    }
+
+}
